Person.cpp
This is a console-based program that manages a database of people and their details such as name, weight, and height. It also calculates the number of calories needed to lose weight based on the person's activity level. The program allows the user to add, remove, update, and display people in the database.

The program starts by displaying a menu with the following options:

Add person
Remove person
Update person
Display all persons
Display calories for person
Exit
When the user selects an option, the program prompts for the required input and performs the corresponding action.

Option 1 (Add person) prompts the user to enter the person's name, last name, weight, and height. It then creates a Person object with the provided details and adds it to the database.

Option 2 (Remove person) prompts the user to enter the name and last name of the person to be removed. It then removes the person from the database.

Option 3 (Update person) prompts the user to enter the name and last name of the person to be updated. It then retrieves the person from the database, prompts the user to enter the updated details, and updates the person in the database.

Option 4 (Display all persons) displays the details of all the people in the database.

Option 5 (Display calories for person) prompts the user to enter the name and last name of the person. It then retrieves the person from the database and calculates the number of calories needed to lose weight based on the person's activity level. It then displays the person's details along with the calculated calories.

Option 6 (Exit) exits the program.